+++ .dotnet.azure/sdk/openai/Azure.AI.OpenAI/tests/FineTuningTests.cs
@@ -338,7 +338,7 @@ public class FineTuningTests : AoaiTestBase<FineTuningClient>
         uploadedFile = await WaitUntilReturnLast(
             uploadedFile,
             () => fileClient.GetFileAsync(uploadedFile.Id),
-            f => f.Status == OpenAIFileStatus.Processed || f.Status == OpenAIFileStatus.Error,
+            f => f.Status == FileStatus.Processed || f.Status == FileStatus.Error,
             TimeSpan.FromSeconds(5),
             TimeSpan.FromMinutes(5))
             .ConfigureAwait(false);